Update 2
So in the MVP stage, I'm not doing git commit messages the nice way. And I'm not splitting them up.

You have to live with that.

But I did create scenes for the game! That's nice, so that I can split up some logic.

I looked at the physis engine, and tried to replace it. But it seems easier just to go with this one. Still need to know how to do colisions, but this is hard it seems.